requirements.txt
streamlit
openai
langchain
langchain-core
python-docx
PyPDF2
pdfplumber